+++ nodejs/build-json.js
@@ -16,10 +16,7 @@ const fs = require("fs"),
 /**
  * Builds the JSON.
  * @param {boolean} clearBuffer if true: the buffer is cleared before the build
- * @returns {number}
- *      0 if fetching the logs fails.
- *      1 if building the json completes successfully.
- *      2 if building wasn't necessary because there are no new log lines.
+ * @returns {number} the according constants.tokens.build_json constant (see constants.js)
  */
 module.exports = function (clearBuffer) {
   if (clearBuffer)
@@ -28,22 +25,15 @@ module.exports = function (clearBuffer) {
   log.info(module, "Starting JSON build.");
   miscFunctions.setProgramStartDate();
 
-  let rebuildRequired = false;
-  switch (fetchLogs()) {
-    case 0:
-      log.info(module, "The build function will now exit!\n");
-      return 0;
-
-    case 1:
-      break;
-
-    case 2:
-      rebuildRequired = true;
+  const fetchLogsReturn = fetchLogs();
+  if (fetchLogsReturn === Constants.tokens.fetch_logs.ERROR_LOG_FETCHING) {
+    log.info(module, "The build function will now exit!\n");
+    return Constants.tokens.build_json.ERROR_LOG_FETCHING;
   }
 
   const lastModification = fs.statSync(data.TS3LogDirectory + data.Logs[data.Logs.length - 1].logName).mtime.valueOf();
 
-  if (rebuildRequired)
+  if (fetchLogsReturn === Constants.tokens.fetch_logs.REBUILD_REQUIRED)
     log.warn(module, "Rebuild required, log order changed or logs were deleted, skipping lastModification check.");
 
   else if (clearBuffer || data.disableLastModificationCheck)
@@ -54,7 +44,7 @@ module.exports = function (clearBuffer) {
 
   else if (lastModification === data.lastModificationOfTheLastLog) {
     log.info(module, "No modifications to the last log since the last request, stopping build process.");
-    return 2;
+    return Constants.tokens.build_json.NOT_NECESSARY;
   }
 
   data.lastModificationOfTheLastLog = lastModification;
@@ -70,5 +60,5 @@ module.exports = function (clearBuffer) {
   miscFunctions.resetConnectedStates();
 
   log.info(module, "Finished JSON build (" + miscFunctions.getProgramRuntime() + " ms).");
-  return 1;
+  return Constants.tokens.build_json.SUCCESS;
 };